+++ .vscode/settings.json
@@ -1,0 +1,10 @@
+{
+    "cSpell.words": [
+        "ENVVAR",
+        "jsonschema",
+        "loginfo",
+        "rospy",
+        "websockets",
+        "wheter"
+    ]
+}
