+++ massrobotics_amr_sender/sample/rosbag/map/README.md
@@ -1,0 +1,3 @@
+# rosbag demo map
+
+This is the map obtained from `turtlebot3` [SLAM simulation](https://emanual.robotis.com/docs/en/platform/turtlebot3/slam_simulation/)
