--- massrobotics_amr_sender_py/.vscode/settings.json
@@ -1,10 +1,0 @@
-{
-    "python.pythonPath": "/usr/bin/python3",
-    "python.testing.pytestArgs": [
-        "test"
-    ],
-    "python.testing.unittestEnabled": false,
-    "python.testing.nosetestsEnabled": false,
-    "python.testing.pytestEnabled": true,
-    "restructuredtext.confPath": ""
-}
