--- massrobotics_amr_sender_py/CONTRIBUTING.md
@@ -1,3 +1,0 @@
-Any contribution that you make to this repository will
-be under the 3-Clause BSD License, as dictated by that
-[license](https://opensource.org/licenses/BSD-3-Clause).
