--- massrobotics_amr_sender_py/package.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-  <name>massrobotics_amr_sender</name>
-  <version>1.0.0</version>
-  <description>MassRobotics AMR Interop Sender</description>
-  <maintainer email="support@inorbit.ai">InOrbit</maintainer>
-  <license>3-Clause BSD License</license>
-
-  <exec_depend>python3-websockets</exec_depend>
-  <!-- kdl is used for processing Twist messages -->
-  <exec_depend>tf2_kdl</exec_depend>
-  <exec_depend>python3-pykdl</exec_depend>
-  <exec_depend>rclpy</exec_depend>
-
-  <test_depend>ament_copyright</test_depend>
-  <test_depend>ament_flake8</test_depend>
-  <test_depend>ament_pep257</test_depend>
-  <!-- common_interfaces for nav_msgs -->
-  <test_depend>common_interfaces</test_depend>
-  <test_depend>python3-jsonschema</test_depend>
-  <test_depend>python3-mock</test_depend>
-  <test_depend>python3-pep8</test_depend>
-  <test_depend>python3-pytest</test_depend>
-  <test_depend>python3-pytest-mock</test_depend>
-  <test_depend>python3-yaml</test_depend>
-
-  <export>
-    <build_type>ament_python</build_type>
-  </export>
-</package>
