--- massrobotics_amr_sender_py/sample/massrobotics_amr_sender_rosbag_launch.py
@@ -1,70 +1,0 @@
-# Copyright 2021 InOrbit, Inc.
-#
-# Redistribution and use in source and binary forms, with or without
-# modification, are permitted provided that the following conditions are met:
-#
-#    * Redistributions of source code must retain the above copyright
-#      notice, this list of conditions and the following disclaimer.
-#
-#    * Redistributions in binary form must reproduce the above copyright
-#      notice, this list of conditions and the following disclaimer in the
-#      documentation and/or other materials provided with the distribution.
-#
-#    * Neither the name of the InOrbit, Inc. nor the names of its
-#      contributors may be used to endorse or promote products derived from
-#      this software without specific prior written permission.
-#
-# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-# AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-# IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
-# ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
-# LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
-# CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
-# SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
-# INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
-# CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
-# ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
-# POSSIBILITY OF SUCH DAMAGE.
-
-
-from launch import LaunchDescription
-from launch.actions import DeclareLaunchArgument, ExecuteProcess
-from launch.substitutions import LaunchConfiguration
-from launch.substitutions import PathJoinSubstitution
-from launch.substitutions import ThisLaunchFileDir
-from launch_ros.actions import Node
-
-
-def generate_launch_description():
-    return LaunchDescription(
-        [
-            DeclareLaunchArgument(
-                "mass_config_file",
-                default_value=[
-                    PathJoinSubstitution([ThisLaunchFileDir(), "config.yaml"])
-                ],
-                description="massrobotics_amr_sender node configuration file",
-            ),
-            DeclareLaunchArgument(
-                "bag_path",
-                default_value=[
-                    PathJoinSubstitution(
-                        [ThisLaunchFileDir(), "rosbag", "rosbag_demo.db3"]
-                    )
-                ],
-                description="Path for ROS 2 data bag",
-            ),
-            Node(
-                package="massrobotics_amr_sender",
-                namespace="massrobotics_amr_sender",
-                executable="massrobotics_amr_node",
-                name="massrobotics_amr_sender",
-                parameters=[{"config_file": LaunchConfiguration("mass_config_file")}],
-            ),
-            ExecuteProcess(
-                cmd=["ros2", "bag", "play", "-l", LaunchConfiguration("bag_path")],
-                output="screen",
-                name="rosbag_demo",
-            ),
-        ]
-    )
