--- massrobotics_amr_sender_py/sample/rosbag/map/map.yaml
@@ -1,7 +1,0 @@
-image: ./map.pgm
-mode: trinary
-resolution: 0.05
-origin: [-1.25, -2.39, 0]
-negate: 0
-occupied_thresh: 0.65
-free_thresh: 0.25
